pyproject.toml
[project]
name = "recruitflow-interviews"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.1",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.4",
    "livekit-agents[google]~=1.0",
    "livekit-api>=1.0.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai>=1.0.20",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-dotenv",
    "reportlab>=4.4.1",
]
